# perf: various performance improvements