--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -980,19 +980,19 @@ module.exports = {
             const parenStack = [];
             const parenPairs = [];
 
-            tokens.forEach(nextToken => {
+            for (let i = 0; i < tokens.length; i++) {
+                const nextToken = tokens[i];
 
-                // Accumulate a list of parenthesis pairs
                 if (astUtils.isOpeningParenToken(nextToken)) {
                     parenStack.push(nextToken);
                 } else if (astUtils.isClosingParenToken(nextToken)) {
-                    parenPairs.unshift({ left: parenStack.pop(), right: nextToken });
+                    parenPairs.push({ left: parenStack.pop(), right: nextToken });
                 }
-            });
+            }
 
-            parenPairs.forEach(pair => {
-                const leftParen = pair.left;
-                const rightParen = pair.right;
+            for (let i = parenPairs.length - 1; i >= 0; i--) {
+                const leftParen = parenPairs[i].left;
+                const rightParen = parenPairs[i].right;
 
                 // We only want to handle parens around expressions, so exclude parentheses that are in function parameters and function call arguments.
                 if (!parameterParens.has(leftParen) && !parameterParens.has(rightParen)) {
@@ -1006,7 +1006,7 @@ module.exports = {
                 }
 
                 offsets.setDesiredOffset(rightParen, leftParen, 0);
-            });
+            }
         }
 
         /**