--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -1731,9 +1731,13 @@ module.exports = {
                     }
 
                     // Invoke the queued offset listeners for the nodes that aren't ignored.
-                    listenerCallQueue
-                        .filter(nodeInfo => !ignoredNodes.has(nodeInfo.node))
-                        .forEach(nodeInfo => nodeInfo.listener(nodeInfo.node));
+                    for (let i = 0; i < listenerCallQueue.length; i++) {
+                        const nodeInfo = listenerCallQueue[i];
+
+                        if (!ignoredNodes.has(nodeInfo.node)) {
+                            nodeInfo.listener(nodeInfo.node);
+                        }
+                    }
 
                     // Update the offsets for ignored nodes to prevent their child tokens from being reported.
                     ignoredNodes.forEach(ignoreNode);
@@ -1744,27 +1748,31 @@ module.exports = {
                      * Create a Map from (tokenOrComment) => (precedingToken).
                      * This is necessary because sourceCode.getTokenBefore does not handle a comment as an argument correctly.
                      */
-                    const precedingTokens = sourceCode.ast.comments.reduce((commentMap, comment) => {
+                    const precedingTokens = new WeakMap();
+
+                    for (let i = 0; i < sourceCode.ast.comments.length; i++) {
+                        const comment = sourceCode.ast.comments[i];
+
                         const tokenOrCommentBefore = sourceCode.getTokenBefore(comment, { includeComments: true });
+                        const hasToken = precedingTokens.has(tokenOrCommentBefore) ? precedingTokens.get(tokenOrCommentBefore) : tokenOrCommentBefore;
 
-                        return commentMap.set(comment, commentMap.has(tokenOrCommentBefore) ? commentMap.get(tokenOrCommentBefore) : tokenOrCommentBefore);
-                    }, new WeakMap());
+                        precedingTokens.set(comment, hasToken);
+                    }
 
-                    sourceCode.lines.forEach((line, lineIndex) => {
-                        const lineNumber = lineIndex + 1;
+                    for (let i = 1; i < sourceCode.lines.length + 1; i++) {
 
-                        if (!tokenInfo.firstTokensByLineNumber.has(lineNumber)) {
+                        if (!tokenInfo.firstTokensByLineNumber.has(i)) {
 
                             // Don't check indentation on blank lines
-                            return;
+                            continue;
                         }
 
-                        const firstTokenOfLine = tokenInfo.firstTokensByLineNumber.get(lineNumber);
+                        const firstTokenOfLine = tokenInfo.firstTokensByLineNumber.get(i);
 
-                        if (firstTokenOfLine.loc.start.line !== lineNumber) {
+                        if (firstTokenOfLine.loc.start.line !== i) {
 
                             // Don't check the indentation of multi-line tokens (e.g. template literals or block comments) twice.
-                            return;
+                            continue;
                         }
 
                         if (astUtils.isCommentToken(firstTokenOfLine)) {
@@ -1789,18 +1797,18 @@ module.exports = {
                                 mayAlignWithBefore && validateTokenIndent(firstTokenOfLine, offsets.getDesiredIndent(tokenBefore)) ||
                                 mayAlignWithAfter && validateTokenIndent(firstTokenOfLine, offsets.getDesiredIndent(tokenAfter))
                             ) {
-                                return;
+                                continue;
                             }
                         }
 
                         // If the token matches the expected indentation, don't report it.
                         if (validateTokenIndent(firstTokenOfLine, offsets.getDesiredIndent(firstTokenOfLine))) {
-                            return;
+                            continue;
                         }
 
                         // Otherwise, report the token/comment.
                         report(firstTokenOfLine, offsets.getDesiredIndent(firstTokenOfLine));
-                    });
+                    }
                 }
             }
         );