--- a/lib/rules/no-loss-of-precision.js
+++ b/lib/rules/no-loss-of-precision.js
@@ -83,7 +83,7 @@ module.exports = {
          * @returns {string} the numeric string with a decimal point in the proper place
          */
         function addDecimalPointToNumber(stringNumber) {
-            return `${stringNumber.slice(0, 1)}.${stringNumber.slice(1)}`;
+            return `${stringNumber[0]}.${stringNumber.slice(1)}`;
         }
 
         /**
@@ -92,7 +92,12 @@ module.exports = {
          * @returns {string} the stripped string
          */
         function removeLeadingZeros(numberAsString) {
-            return numberAsString.replace(/^0*/u, "");
+            for (let i = 0; i < numberAsString.length; i++) {
+                if (numberAsString[i] !== "0") {
+                    return numberAsString.slice(i);
+                }
+            }
+            return numberAsString;
         }
 
         /**
@@ -101,7 +106,12 @@ module.exports = {
          * @returns {string} the stripped string
          */
         function removeTrailingZeros(numberAsString) {
-            return numberAsString.replace(/0*$/u, "");
+            for (let i = numberAsString.length - 1; i >= 0; i--) {
+                if (numberAsString[i] !== "0") {
+                    return numberAsString.slice(0, i + 1);
+                }
+            }
+            return numberAsString;
         }
 
         /**