--- a/lib/rules/no-loss-of-precision.js
+++ b/lib/rules/no-loss-of-precision.js
@@ -138,7 +138,7 @@ module.exports = {
             const trimmedFloat = removeLeadingZeros(stringFloat);
 
             if (trimmedFloat.startsWith(".")) {
-                const decimalDigits = trimmedFloat.split(".").pop();
+                const decimalDigits = trimmedFloat.slice(1);
                 const significantDigits = removeLeadingZeros(decimalDigits);
 
                 return {
@@ -153,7 +153,6 @@ module.exports = {
 
             };
         }
-
 
         /**
          * Converts a base ten number to proper scientific notation
@@ -170,7 +169,6 @@ module.exports = {
                 : normalizedNumber.magnitude;
 
             return `${normalizedCoefficient}e${magnitude}`;
-
         }
 
         /**