@mdjermanovic 2023-04-29
> overhead from the functional tools (map, filter, reduce)

These methods are commonly used, and therefore highly optimized in JS engines. I wouldn't expect for-loops to be noticeably faster (in fact, it might be the opposite), and that might depend on the engine and its version, so unless there's a significant difference in performance, I'm not in favor of making these changes.