@moonlightaria 2023-04-30
[benchs.zip](https://github.com/eslint/eslint/files/11362011/benchs.zip)
here are the benchmarks on before and after the changes i just ran

for foreach it is often faster for very small collections since there is less code for js to compile but filter and map always creates a new array which becomes a performance burden expecially when chained since each on creates it own array