@moonlightaria 2023-05-09
benchmark command: node --cpu-prof ../../bin/eslint.js large.js --no-ignore
decided to run new benchmarks and several for better info, lost benchmark command used during initial runs so used new ones

some improvements are minimal while other are major, optimizations which reduce memory allocation remove gc cycles which can occur outside of the functions which isn't captured in this quick summery

if functions didn't appear it means no trace was grabbed during the execution of that function generally meaning that execution time to too small matter

addparensindent:
old:
438.54 ms
411.72 ms
459.63 ms
avg: 436.63 ms

new:
433.95 ms
437.45 ms
429.41 ms
avg: 433.60 ms 

removeLeadingZeros:
old:
30.34 ms
20.03 ms
27.29 ms

new:
didn't appear
didn't appear
1.26 ms

removeTrailingZeros:
old:
8.04 ms
22.84 ms
11.05 ms

new:
didn't appear
didn't appear
2.61 ms

checklist:
old:
3.93 ms
.76 ms
1.24 ms
avg: 1.97 ms

new:
.42 ms
.69 ms
2.13 ms
avg: 1.08 ms

calculateStatsPerFile:
old:
26.82 ms
8.87 ms
23.44 ms
avg: 19.71 ms

new:
4.82 ms
4.88 ms
9.84 ms
avg: 6.51 ms

calculateStatsPerRun:
no appearance on any profile, same optimization as calculateStatsPerFile

checkProgramForMaxLength:
old:
168.78 ms
148.38 ms
178.24 ms
avg: 165.13 ms

new:
157.43 ms
153.75 ms
151.78 ms
avg: 154.32 ms


TokenInfo constructor:
old:
34.92 ms
39.39 ms
38.49 ms
avg: 37.6

new:
24.88 ms
34.23 ms
25.47 ms
avg: 28.19

check:
no appearance on any profile, replaced filter array creation and length check with incrementing an integer

create:
old:
37.75 ms
41.10 ms
33.16 ms
avg: 37.33 ms

new:
37.12 ms
39.41 ms
41.81 ms
avg: 39.44 ms