@nzakas 2023-06-29
@moonlightaria we'd like to pay you for this contribution. Please contact us at contact (at) eslint (dot) org.